+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_transports.php
@@ -3,7 +3,6 @@
 $container->loadFromExtension('framework', array(
     'serializer' => true,
     'messenger' => array(
-        'serializer' => true,
         'transports' => array(
             'default' => 'amqp://localhost/%2f/messages',
             'customised' => array(