+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/serializer_disabled.php
@@ -4,7 +4,4 @@ $container->loadFromExtension('framework', array(
     'serializer' => array(
         'enabled' => false,
     ),
-    'messenger' => array(
-        'serializer' => false,
-    ),
 ));