+++ src/Symfony/Component/Messenger/Tests/Transport/AmqpExt/AmqpTransportFactoryTest.php
@@ -15,16 +15,14 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Transport\AmqpExt\AmqpTransport;
 use Symfony\Component\Messenger\Transport\AmqpExt\AmqpTransportFactory;
 use Symfony\Component\Messenger\Transport\AmqpExt\Connection;
-use Symfony\Component\Messenger\Transport\Serialization\DecoderInterface;
-use Symfony\Component\Messenger\Transport\Serialization\EncoderInterface;
+use Symfony\Component\Messenger\Transport\Serialization\SerializerInterface;
 
 class AmqpTransportFactoryTest extends TestCase
 {
     public function testSupportsOnlyAmqpTransports()
     {
         $factory = new AmqpTransportFactory(
-            $this->getMockBuilder(EncoderInterface::class)->getMock(),
-            $this->getMockBuilder(DecoderInterface::class)->getMock(),
+            $this->getMockBuilder(SerializerInterface::class)->getMock(),
             true
         );
 
@@ -36,12 +34,11 @@ class AmqpTransportFactoryTest extends TestCase
     public function testItCreatesTheTransport()
     {
         $factory = new AmqpTransportFactory(
-            $encoder = $this->getMockBuilder(EncoderInterface::class)->getMock(),
-            $decoder = $this->getMockBuilder(DecoderInterface::class)->getMock(),
+            $serializer = $this->getMockBuilder(SerializerInterface::class)->getMock(),
             true
         );
 
-        $expectedTransport = new AmqpTransport($encoder, $decoder, Connection::fromDsn('amqp://localhost', array('foo' => 'bar'), true), array('foo' => 'bar'), true);
+        $expectedTransport = new AmqpTransport($serializer, Connection::fromDsn('amqp://localhost', array('foo' => 'bar'), true));
 
         $this->assertEquals($expectedTransport, $factory->createTransport('amqp://localhost', array('foo' => 'bar')));
     }