+++ src/Symfony/Component/Messenger/Transport/Serialization/SerializerInterface.php
@@ -16,17 +16,27 @@ use Symfony\Component\Messenger\Envelope;
 /**
  * @author Samuel Roze <samuel.roze@gmail.com>
  */
-interface EncoderInterface
+interface SerializerInterface
 {
     /**
+     * Decodes an envelope and its message from an encoded-form.
+     *
+     * The `$encodedEnvelope` parameter is a key-value array that
+     * describes the envelope and its content, that will be used by the different transports.
+     *
+     * The most common keys are:
+     * - `body` (string) - the message body
+     * - `headers` (string<string>) - a key/value pair of headers
+     */
+    public function decode(array $encodedEnvelope): Envelope;
+
+    /**
      * Encodes an envelope content (message & items) to a common format understandable by transports.
-     * The encoded array should only contain scalar and arrays.
+     * The encoded array should only contain scalars and arrays.
      *
      * The most common keys of the encoded array are:
      * - `body` (string) - the message body
      * - `headers` (string<string>) - a key/value pair of headers
-     *
-     * @param Envelope $envelope The envelop containing the message put on the MessageBus by the user
      */
     public function encode(Envelope $envelope): array;
 }